Updated channel name, and set a display and tag
The custom channel name adds a Jira project key to the channel name.
The display setting adds a link to the Jira issue in the title of the issue (only when displayed, not edited).
The tag setting tells Conclude to display a link to the Jira issue in Conclude Inbox.