Prefix default activity channel name with "_"
Use the name of the install channel as the basis of creating activity channels. For example, if the app is installed in the channel support, channelname will become _support_${jira_key} and each support ticket will have the Jira key in the channel name. If the Jira key is CU, the channels will be _support_cu-1, *cu-2 etc.